Chore/resolve the problem in the review

+++ src/main/java/org/cardanofoundation/explorer/api/projection/LatestVotingProcedureProjection.java
@@ -28,7 +28,7 @@ public interface LatestVotingProcedureProjection {
 
   Integer getBlock();
 
-  BigInteger getBlockTime();
+  Long getBlockTime();
 
   Boolean getRepeatVote();
